+++ payments-core/src/main/java/com/stripe/android/googlepaylauncher/injection/GooglePayLauncherModule.kt
@@ -28,7 +28,8 @@ class GooglePayLauncherModule {
             environment,
             GooglePayJsonFactory.BillingAddressParameters(),
             true,
-            logger
+            logger,
+            true
         )
     }
 }